Remove duplication in object prims
